# api/rest/ping_api.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "ping_api.proto",
    "version": "version not set"
  },
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/v1/pong/stream": {
      "post": {
        "summary": "PongStream accepts a request stating how many requests to stream back.\nIt will response with a stream until the count is reached and close the \nconnection.",
        "operationId": "PongStream",
        "responses": {
          "200": {
            "description": "A successful response.(streaming responses)",
            "schema": {
              "$ref": "#/x-stream-definitions/v1PongStreamResponse"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1PongStreamRequest"
            }
          }
        ],
        "tags": [
          "PingAPI"
        ]
      }
    }
  },
  "definitions": {
...
    "v1PongStreamRequest": {
      "type": "object",
      "properties": {
        "count": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        }
      },
      "title": "PongStreamRequest contains a count telling the server how many responses\nto stream, and a message to stream back"
    },
...
  },
...
}